Fix issue where user preferences for check-in intervals are not saved
Update interval hours to use parseFloat to correctly store and retrieve decimal values, resolving an issue where the setting would revert to the default.

+++ server/storage.ts
@@ -354,7 +354,7 @@ class DatabaseStorage implements IStorage {
     const row = result[0];
     return {
       frequency: row.frequency as "daily" | "every_two_days",
-      intervalHours: parseInt(row.intervalHours) || 24,
+      intervalHours: parseFloat(row.intervalHours) || 24,
       lastCheckIn: row.lastCheckIn?.toISOString() || null,
       nextCheckInDue: row.nextCheckInDue?.toISOString() || null,
       alertsEnabled: row.alertsEnabled,
